#90 refactor: profilePage 컴포넌트 리팩토링
