+++ src/app/profile/page.tsx
@@ -2,11 +2,6 @@ import { ProfilePageComponent } from '@/components/Profile';
 import { getUserCookies } from '@/shared/@common/utils/getUserCookies';
 
 export default function Profile() {
-  const { loginedId, accessToken } = getUserCookies();
-  return (
-    <ProfilePageComponent
-      loginedId={Number(loginedId)}
-      accessToken={accessToken}
-    />
-  );
+  const { loginedId } = getUserCookies();
+  return <ProfilePageComponent userId={Number(loginedId)} />;
 }
